Create FibonacciSumLastDigit.java
Given an integer n, find the last digit of the sum F(0) + F(1) + ... + F(n)
F(n) is the nth Fibonacci number
n can be as large as 10^14

Iterating over all the Fibonacci numbers fails (overflow).
Iterating to n by keeping only the last digit of each Fibonacci number and the running total (as I do in the solution) is too slow.

We were given a hint to print the "first few" last digits of sums.
After printing ~100 digits, it's apparent that the sequence is periodic.
Here are the last digits of the first 120 sums:

0 1 2 4 7 2 0 3 4 8 3 2 6 9 6 6 3 0 4 5 0 6 7 4 2 7 0 8 9 8 8 7 6 4 1 6 8 5 4 0 5 6 2 9 2 2 5 8 4 3 8 2 1 4 6 1 8 0 9 0
0 1 2 4 7 2 0 3 4 8 3 2 6 9 6 6 3 0 4 5 0 6 7 4 2 7 0 8 9 8 8 7 6 4 1 6 8 5 4 0 5 6 2 9 2 2 5 8 4 3 8 2 1 4 6 1 8 0 9 0

The period length is 60. This period, like the Pisano, begins with 0 1.

The period length is the key: to find the last digit of the sum of Fibonacci numbers from F(0) to F(n) for any n, you need only find the last digit of the sum up to F(n mod 60).