Create LargestNumber.java
Introduction:
As the last question of a successful interview, your boss gives you a few pieces of paper with numbers on them and asks you to compose the largest number from these numbers. The resulting number is going to be your salary, so you are very interested in maximizing this number.